+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 ## Improvements
 
 - [OData] Add warning for deep update attempts in OData v2. 
+- [Generator] The transpiling after generation is done in chunks to avoid CPU overloads for high numbers of services.
 
 ## Fixed Issues